+++ examples/exidx_unwinder_test.cpp
@@ -155,7 +155,8 @@ int main() {
   auto *stack_region = reinterpret_cast<uint32_t *>(&target_memory[stack_base]);
 
   // Frame 0 (SensorData_Process active frame)
-  stack_region[0] = stack_base + 16; // Caller FP (points to Frame 1)
+  stack_region[0] =
+      static_cast<uint32_t>(stack_base + 16); // Caller FP (points to Frame 1)
   stack_region[1] = 0x0800'2011;     // Thumb return PC in System_MainLoop
 
   // Frame 1 (System_MainLoop caller frame)
